data_components/data_augmentation/requirements.txt
pandas
pandas_market_calendars
tqdm
PyWavelets
matplotlib
plotly
IPython
gcsfs
ta
google-cloud-pipeline-components
kaleido